Extracted menu functionality into a common superclass
