+++ src/main/java/ch/bbbaden/idpa/bru_eap_mey/quiz/controllers/CategoryEditController.java
@@ -1,36 +1,8 @@
 package ch.bbbaden.idpa.bru_eap_mey.quiz.controllers;
 
-import org.eclipse.jdt.annotation.Nullable;
-
-
-import ch.bbbaden.idpa.bru_eap_mey.quiz.model.QuizModel;
-
 /**
  * Der Controller für die Übersicht der Kategorien.
  */
-public class CategoryEditController {
-	
-	/**
-	 * Das Quizmodel
-	 */
-	private @Nullable QuizModel quizModel;
-	
-	/**
-	 * Initialisiert das Quizmodel.
-	 * 
-	 * @param model
-	 *        das Model
-	 */
-	public final void setModel(QuizModel model) {
-		this.quizModel = model;
-	}
-	
-	/**
-	 * @return
-	 * 		das Quizmodel dieses Controllers
-	 */
-	public final QuizModel getModel() {
-		assert this.quizModel != null;
-		return this.quizModel;
-	}
+public class CategoryEditController extends MainMenuController {
+	//
 }
